ISSUE-76: Refactor code to use common arrows fragment
