+++ app/src/main/java/ar/com/tzulberti/archerytraining/model/base/ISerie.java
@@ -1,12 +1,21 @@
 package ar.com.tzulberti.archerytraining.model.base;
 
+import java.io.Serializable;
 import java.util.List;
 
 /**
  * Created by tzulberti on 6/3/17.
  */
 
-public interface ISerie {
+public interface ISerie extends Serializable {
 
-    public List<AbstractArrow> getArrows();
+    List<? extends AbstractArrow> getArrows();
+
+    int getIndex();
+
+    ISerieContainer getContainer();
+
+    int getTotalScore();
+
+    void updateTotalScore(int arrowScore);
 }
