+++ app/src/main/java/ar/com/tzulberti/archerytraining/model/base/ISerieContainer.java
@@ -1,0 +1,14 @@
+package ar.com.tzulberti.archerytraining.model.base;
+
+import java.util.List;
+
+/**
+ * Created by tzulberti on 6/3/17.
+ */
+
+public interface ISerieContainer {
+
+    List<? extends ISerie> getSeries();
+
+    long getId();
+}
