+++ app/src/main/java/ar/com/tzulberti/archerytraining/model/tournament/TournamentConfiguration.java
@@ -1,10 +1,12 @@
 package ar.com.tzulberti.archerytraining.model.tournament;
 
+import java.io.Serializable;
+
 /**
  * Created by tzulberti on 5/22/17.
  */
 
-public class TournamentConfiguration {
+public class TournamentConfiguration implements Serializable {
 
     public static final int MAX_SERIES = 12;
     public static final int MAX_ARROW_PER_SERIES = 6;
