+++ app/src/main/java/ar/com/tzulberti/archerytraining/model/tournament/TournamentSerie.java
@@ -1,13 +1,18 @@
 package ar.com.tzulberti.archerytraining.model.tournament;
 
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
+
+import ar.com.tzulberti.archerytraining.model.base.AbstractArrow;
+import ar.com.tzulberti.archerytraining.model.base.ISerie;
+import ar.com.tzulberti.archerytraining.model.base.ISerieContainer;
 
 /**
  * Created by tzulberti on 5/22/17.
  */
 
-public class TournamentSerie {
+public class TournamentSerie implements ISerie, Serializable {
 
     public long id;
     public int index;
@@ -28,5 +33,30 @@ public class TournamentSerie {
         } else {
             return false;
         }
+    }
+
+    @Override
+    public List<? extends AbstractArrow> getArrows() {
+        return this.arrows;
+    }
+
+    @Override
+    public int getIndex() {
+        return this.index;
+    }
+
+    @Override
+    public ISerieContainer getContainer() {
+        return this.tournament;
+    }
+
+    @Override
+    public int getTotalScore() {
+        return totalScore;
+    }
+
+    @Override
+    public void updateTotalScore(int arrowScore) {
+        this.totalScore += arrowScore;
     }
 }
